requirements.txt
numpy
pandas
scikit-learn
matplotlib
keras
joblib
streamlit
pillow
tensorflow-cpu
pickle-mixin
nltk
PyYAML